Fix google calendar embed
The /b/1/ in the URL meant that if you were logged into 2 google accounts, the 1st account (not 0th) required access.
If we try /b/0 then it only works if the 0th account has access.

Without /b/N/ in the URL it appears to try all accounts - i.e. it works with my shef.ac.uk account as either 0th or 1st.

Fixes #103