R Resources blog post
I had to provide some R resources to someone for a new project. Recently a collaborator was also having to start out
with learning R and I shared some of my links with them.

Figured it would be useful to make a blog post out of it so we've a resource to point people to in the future. It can of
course be updated as new resources come online.

I've checked the majority of links should work (found a few dated ones for books).

Feedback more than welcome, too much, too many sections, not enough focus?

@ubdbra001 you know more about R than me so if there is any glaring omission please let me know but if there are
significant additional resources to be added I'd suggest making a separate issue/PR after this is published.