Fixes error in code block which messed up rendering
This article started live in my [org-roam](https://www.orgroam.com/) notes and was exported to Markdown. In the export
it retained the code chunk options of whether to evaluate code blocks which Markdown doesn't support and I missed
removing one of them when publishing the article.

This removes it and should fix rendering of the relevant section.