blog : pytest parametrisation
A blog post I developed on my [personal blog](https://ns-rse.github.io/posts/pytest-param/) (as I often do) that I
thought would be useful to have on the RSE blog too.

One thing I'm unsure about is whether the `social_image` target can be external to the site. Whilst ultimately
superfluous I think pictures are nice (this one is of various stones and minerals on a table top in the National Natural
History Museum in Ljubjiana and I thought it was appropriate as its lots of similar things but slightly different, like
the concept of parametrisation :-)