pages: call for proposals 2025
Closes #897

Moves the 2024 page to a new nested location and adds a 2025 page.

The 2025 page is commented out of `_data/navigation.yml` and so _doesn't_ appear in the rendered pages. To view navigate to the 2024 page `Ctrl + l` to get to the navigation bar and change `/2024/` to `/2025/` and hit `Ret`.

When page is ready to go live this will be uncommented and will also require adding to the front-page. Will address in a separate issue/pull-request.