newsletter: November 2025
Closes #870

HI version of the newsletter. Took an hour to go through the links, remove outdated, check details, copy/paste/edit/lint.

I'll give the AI-bot a whirl later when I've time to allow my memory of what the details are to fade and see how long it
takes to proof and tweak for comparisson.